Add files via upload
Projects from PyCharm (HyperSkill) Rules for Rock Paper Scissors Game !!

Whatever list of options the user chooses, your program, obviously, should be able to identify which option beats which, that is, the relationships between different options. First, every option is equal to itself (causing a draw if both the user and the computer choose the same option). Secondly, every option wins over one half of the other options of the list and gets defeated by another half. How to determine which options are stronger or weaker than the option you're currently looking at? Well, you can try to do it this way: take the list of options (provided by the user or the default one). Find the option for which you want to know its relationships with other options. Take all the options that follow this chosen option in the list. Add to them the list of options that precede the chosen option. Now you have another list of options that doesn't include the "chosen" option and has a different order of elements in it (first go the options following the chosen one in the original list, after them are the ones that precede it). So, in this "new" list, the first half of the options will be defeating the "chosen" option, and the second half will get beaten by it.

For example, the user's list of options is rock,paper,scissors,lizard,spock. You want to know what options are weaker than lizard. By looking at the list spock,rock,paper,scissors you realize that spock and rock will be beating the lizard, and paper and scissors will get defeated by it. For spock it'll be almost the same, but it'll get beaten by rock and paper, and prevail over scissors and lizard. For the version of the game with 15 options, you can look at the picture above to understand the relationships between options.

Your program should:

Output a line Enter your name: . Note that the user should enter his/her name on the same line (not the one following the output!) Read input specifying the user's name and output a new line Hello, Read a file named rating.txt and check if there's a record for the user with the same name; if yes, use the score specified in rating.txt for this user as a starting point for calculating the score in the current game. If no, start counting the user's score from 0. Read input specifying the list of options that will be used for playing the game (options are separated by comma). If the input is an empty line, play with default options. Output a line Okay, let's start Play the game by the rules defined in the previous stages: Read the user's input If the input is !exit, output Bye! and stop the game If the input is the name of the option, then: Pick a random option Output a line with the result of the game in the following format ( is the name of the option chosen by the program): Lose -> Sorry, but the computer chose Draw -> There is a draw () Win -> Well done. The computer chose and failed For each draw, add 50 points to the score. For each user's win, add 100 to his/her score. In case the user loses, don't change the score. If the input corresponds to anything else, output Invalid input Play the game again (with the same options that were defined before the start of the game)